+++ src/app/(shared)/about/page.tsx
@@ -1,7 +1,4 @@
-import ServiceFeature from '@/components/public/ServiceFeature';
-import UserMenuFeature from '@/components/public/UserMenuFeature';
-import EditorFeature from '@/components/public/EditorFeature';
-import { Container, Space } from '@mantine/core';
+import AboutPageContent from '@/components/pageContent/AboutPageContent';
 import type { Metadata } from 'next';
 
 export const metadata: Metadata = {
@@ -11,13 +8,5 @@ export const metadata: Metadata = {
 };
 
 export default function Page() {
-  return (
-    <Container my="md">
-      <ServiceFeature />
-      <Space h={60} />
-      <UserMenuFeature />
-      <Space h={20} />
-      <EditorFeature />
-    </Container>
-  );
+  return <AboutPageContent />;
 }
